+++ packages/sdk/src/RpcStreamManager.ts
@@ -13,11 +13,11 @@ import { Readable, type Stream, Writable } from "node:stream";
 import { pipeline } from "node:stream/promises";
 import { Logger } from "@zowe/imperative";
 import { Base64Decode, Base64Encode } from "base64-stream";
-import type { Client, ClientChannel } from "ssh2";
 import type { CallbackInfo, CommandResponse, RpcNotification, RpcPromise, RpcRequest } from "./doc";
 import type { ReadDatasetRequest } from "./doc/rpc/ds";
 import type { ReadFileRequest } from "./doc/rpc/uss";
 import { ProgressTransform } from "./ProgressTransform";
+import type { Client, ClientChannel } from "./ssh-rs";
 
 type StreamData = { stream: Stream; callbackInfo?: CallbackInfo; resourceName?: string };
 type StreamMode = "GET" | "PUT";