+++ packages/sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added support for USS Move operations. [#789](https://github.com/zowe/zowe-native-proto/pull/789)
 - Fixed `user` property not defined as secure when new team configuration profile is created. [#845](https://github.com/zowe/zowe-native-proto/pull/845)
 - Fixed check for outdated server always returning true which could trigger unnecessary re-deploy. [#862](https://github.com/zowe/zowe-native-proto/pull/862)
+- Added experimental native client for improved performance. To enable it, pass the `useNativeSsh` option to the `ZSshClient.create` method. [#833](https://github.com/zowe/zowe-native-proto/pull/833)
 
 ## `0.3.0`